refactor: Modal storybook 구조 변경
