+++ components/src/lib/Modal/storybook/Modal.size.stories.tsx
@@ -1,8 +1,8 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import Modal from './Modal';
+import Modal from '../Modal';
 
 const meta = {
-  title: 'Modal',
+  title: 'Modal/size',
   component: Modal,
 } satisfies Meta<typeof Modal>;
 
@@ -10,7 +10,7 @@ export default meta;
 
 type Story = StoryObj<typeof meta>;
 
-export const Default: Story = {
+export const Small: Story = {
   args: {
     isOpened: true,
     onClose: () => {
@@ -33,39 +33,15 @@ export const Default: Story = {
   },
 };
 
-export const CloseButton: Story = {
+export const Medium: Story = {
   args: {
+    size: 'medium',
     isOpened: true,
     onClose: () => {
       alert('모달 닫기');
     },
     title: '제목입니다',
     description: '설명입니다',
-    primaryButton: {
-      text: '확인',
-      onClick: () => {
-        alert('확인');
-      },
-    },
-    secondaryButton: {
-      text: '취소',
-      onClick: () => {
-        alert('취소');
-      },
-    },
-    showCloseButton: true,
-  },
-};
-
-export const Bottom: Story = {
-  args: {
-    isOpened: true,
-    onClose: () => {
-      alert('모달 닫기');
-    },
-    title: '제목입니다',
-    description: '설명입니다',
-    modalPosition: 'bottom',
     primaryButton: {
       text: '확인',
       onClick: () => {
@@ -81,15 +57,15 @@ export const Bottom: Story = {
   },
 };
 
-export const Column: Story = {
+export const Large: Story = {
   args: {
+    size: 'large',
     isOpened: true,
     onClose: () => {
       alert('모달 닫기');
     },
     title: '제목입니다',
     description: '설명입니다',
-    modalPosition: 'bottom',
     primaryButton: {
       text: '확인',
       onClick: () => {
@@ -102,32 +78,5 @@ export const Column: Story = {
         alert('취소');
       },
     },
-    buttonPosition: 'column',
-  },
-};
-
-export const Color: Story = {
-  args: {
-    isOpened: true,
-    onClose: () => {
-      alert('모달 닫기');
-    },
-    title: '제목입니다',
-    description: '설명입니다',
-    modalPosition: 'bottom',
-    primaryButton: {
-      text: '확인',
-      onClick: () => {
-        alert('확인');
-      },
-    },
-    secondaryButton: {
-      text: '취소',
-      onClick: () => {
-        alert('취소');
-      },
-    },
-    buttonPosition: 'column',
-    primaryColor: '#f66f00',
   },
 };
